Extend elbe init with shared/cloned directories, project repository, apt-cacher, image conversion.
* Shared Directories

To enable shared directories between host and VM also during elbe initialisation
a 2 stage initialisation process is needed. The debian installer does not support
to share a directory by virtio.

On 1st stage the basic debian system is installed in the VM.
On 2nd stage elbe is installed. The shared directory is enabled on this
stage and the project repository in the shared directory may be used to
access any debian packages preloaded by elbe init (e.g. elbe itself).

The two stage install process is commanded by init-elbe.sh. This script
is executed by the debian installer as preseed/late-command. The
script installs itself as an init script, which is executed on the
2nd stage boot to finish the install process.

The shared directories can be defined in the elbe xml file by <initvm/share-list>.
Within the initvm and build environment a directory can be accessed by /media/[id].

* Cloned Directories

Directories are cloned on the 2nd stage boot step.

The directories to be cloned can be defined in the elbe xml file by <initvm/clone-list>.

* Project Repository

Elbe init adds a project packet repository as a shared directory on the 2nd stage
installation step.
The project packet directory can be automatically filled by elbe init.
The project package directory gets the highest priority. Packages
will be searched here first - during VM installation and later on.
Packages to be added can either be specified on the command line or
in the elbe xml file.

The shared project repository can be defined in the elbe xml file by <initvm/repository>
Within the initvm and build environment the repository can be accessed by /media/elbe-repo.

* apt-cacher

An apt-cacher server can be specified to speed up package retrieval. The
apt-cacher will already be used in the 1st stage. It can be defined for the initvm.

* Image Conversion

The build image can be converted to a virtualbox image or a vmware image.
Use make buildimage.vdi or make buildimage.vmdk.
*